pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
boltz = ["boltz", "cuequivariance-torch", "cuequivariance-ops-torch-cu12", "rdkit>=2025.3.6"]
dev = ["pytest", "pytest-cov", "mypy", "pre-commit", "pyright"]
protenix = ["protenix>=0.6.3", "einx", "triton"]

[project]
authors = [{email = "karson.chrispens@ucsf.edu", name = "Karson Chrispens"}]
dependencies = ["atomworks[ml]", "dotenv", "jaxtyping", "jax"]
name = "sampleworks"
requires-python = ">= 3.11, <3.14"
version = "0.1.0"

[tool.pixi.dependencies]
gcc_linux-64 = ">=9,<13"
gxx_linux-64 = ">=9,<13"
ninja = "*"
numpy = "<2.0"
pyarrow = "==17.0.0"

[tool.pixi.environments]
boltz = {features = ["boltz"]}
boltz-dev = {features = ["boltz", "dev"]}
protenix = {features = ["protenix"]}
protenix-dev = {features = ["protenix", "dev"]}

[tool.pixi.feature.protenix.activation.env]
PYTHONNOUSERSITE = "1"

[tool.pixi.feature.protenix.pypi-options.dependency-overrides]
rdkit = ">=2024.3.5,<2025.9"

[tool.pixi.pypi-dependencies]
sampleworks = {editable = true, path = "."}

[tool.pixi.system-requirements]
cuda = "12"

[tool.pixi.tasks]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pytest.ini_options]
addopts = "-v --strict-markers"
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
python_classes = ["Test*"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
testpaths = ["tests"]

[tool.ruff]
extend-include = ["*.ipynb"]
src = []

[tool.ruff.lint]
fixable = ["I001", "F401", "UP"]
ignore = ["E402", "E721", "E731", "E741", "F722"]
select = ["E", "F", "I001", "UP"]

[tool.ruff.lint.flake8-import-conventions.extend-aliases]
"collections" = "co"
"functools" = "ft"
"itertools" = "it"

[tool.ruff.lint.isort]
combine-as-imports = true
extra-standard-library = ["typing_extensions"]
lines-after-imports = 2
order-by-type = false
